gen-bundle: Better support for b1 and refactoring
Non-behavioral changes:
- Move fromHar() and its helpers to separate file (fromhar.go)
- Now main() is responsible for writing output file

Behavioral changes:
- -startURL flag is renamed to -primaryURL and it's now required
- -startURL and -manifestURL must be absolute
