+++ go/bundle/cmd/gen-bundle/fromdir.go
@@ -11,52 +11,9 @@ import (
 	"strings"
 
 	"github.com/WICG/webpackage/go/bundle"
-	"github.com/WICG/webpackage/go/bundle/version"
 )
 
-func fromDir(dir string, ver version.Version, baseURL string, startURL string, manifestURL string) error {
-	parsedBaseURL, err := url.Parse(baseURL)
-	if err != nil {
-		return fmt.Errorf("Failed to parse base URL. err: %v", err)
-	}
-	parsedStartURL, err := parsedBaseURL.Parse(startURL)
-	if err != nil {
-		return fmt.Errorf("Failed to parse start URL. err: %v", err)
-	}
-	var parsedManifestURL *url.URL
-	if len(manifestURL) > 0 {
-		parsedManifestURL, err = parsedBaseURL.Parse(manifestURL)
-		if err != nil {
-			return fmt.Errorf("Failed to parse manifest URL. err: %v", err)
-		}
-	}
-
-	fo, err := os.OpenFile(*flagOutput, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0644)
-	if err != nil {
-		return fmt.Errorf("Failed to open output file %q for writing. err: %v", *flagOutput, err)
-	}
-	defer fo.Close()
-
-	es, err := createExchangesFromDir(dir, parsedBaseURL)
-	if err != nil {
-		return err
-	}
-	b := &bundle.Bundle{Version: ver, PrimaryURL: parsedStartURL, Exchanges: es, ManifestURL: parsedManifestURL}
-	// Move the startURL entry to first.
-	for i, e := range b.Exchanges {
-		if e.Request.URL.String() == parsedStartURL.String() {
-			b.Exchanges[0], b.Exchanges[i] = b.Exchanges[i], b.Exchanges[0]
-			break
-		}
-	}
-
-	if _, err := b.WriteTo(fo); err != nil {
-		return fmt.Errorf("Failed to write exchange. err: %v", err)
-	}
-	return nil
-}
-
-func createExchangesFromDir(baseDir string, baseURL *url.URL) ([]*bundle.Exchange, error) {
+func fromDir(baseDir string, baseURL *url.URL) ([]*bundle.Exchange, error) {
 	es := []*bundle.Exchange{}
 	err := filepath.Walk(baseDir, func(path string, info os.FileInfo, err error) error {
 		if err != nil {
